Rename t3c to t3c-apply
