+++ cache-config/t3c-apply/README.md
@@ -1,6 +1,6 @@
-# t3c
+# t3c-apply
 
-t3c is a transliteration of traffic_ops_ort.pl script to the go language.
+t3c-apply is a transliteration of traffic_ops_ort.pl script to the go language.
 It is designed to replace the traffic_ops_ort.pl perl script and it is used to apply 
 configuration from Traffic Control, stored in Traffic Ops, to the cache.
 
@@ -10,7 +10,7 @@ Typical usage is to install `t3c` on the cache machine, and then run it periodic
 
 ## Options
 
-The `t3c` app has the following command-line options:
+The `t3c-apply` app has the following command-line options:
 
 
 long option                             | short | default | description
@@ -40,13 +40,13 @@ long option                             | short | default | description
 
 # Modes
 
-The `t3c` app can be run in a number of modes.
+The `t3c-apply` app can be run in a number of modes.
 
 The syncds mode is the normal mode of operation, which should typically be run periodically via cron or a similar tool.
 
-The badass mode is typically an emergency-fix mode, which will override and replace all files with the configuration generated from the current Traffic Ops data, regardless whether `t3c` (presumably incorrectly) thinks the files need updating or not. It is recommended to run this mode when something goes wrong, and the configuration on the cache is incorrect, and the data in Traffic Ops and config generation is believed to be correct. It is not recommended to run this in normal operation; use syncds mode for normal operation.
+The badass mode is typically an emergency-fix mode, which will override and replace all files with the configuration generated from the current Traffic Ops data, regardless whether `t3c-apply` (presumably incorrectly) thinks the files need updating or not. It is recommended to run this mode when something goes wrong, and the configuration on the cache is incorrect, and the data in Traffic Ops and config generation is believed to be correct. It is not recommended to run this in normal operation; use syncds mode for normal operation.
 
-The revalidate mode will apply Revalidations from Traffic Ops (regex_revalidate.config) but no other configuration. This mode was intended to quickly apply revalidations when `t3c` took a long time to run. It is less relevant with the current speed of `t3c` but may still be useful on slow networks or very large deployments.
+The revalidate mode will apply Revalidations from Traffic Ops (regex_revalidate.config) but no other configuration. This mode was intended to quickly apply revalidations when `t3c-apply` took a long time to run. It is less relevant with the current speed of `t3c-apply` but may still be useful on slow networks or very large deployments.
 
 mode        | description
 ------------| ---
@@ -57,7 +57,7 @@ revalidate  | checks for updated revalidations in Traffic Ops and applies them
 
 # Behavior
 
-When `t3c` is run, it will:
+When `t3c-apply` is run, it will:
 
 1. Delete all of its temporary directories over a week old. Currently, the base temp directory is hard-coded to /tmp/ort.
 1. Determine if Updates have been Queued on the server (by checking the Server's Update Pending or Revalidate Pending flag in Traffic Ops).
@@ -73,15 +73,15 @@ When `t3c` is run, it will:
     1. **NOTE** the default profiles distributed by Traffic Control have an ATS chkconfig with a runlevel before networking is enabled, which is likely incorrect.
     1. **NOTE** this is not used by CentOS 7+ and ATS 7+. SystemD does not use chkconfig, and ATS 7+ uses a SystemD script not an init script.
 1. Process each config file
-    1. If `t3c` is in revalidate mode, this will only be regex_revalidate.config
+    1. If `t3c-apply` is in revalidate mode, this will only be regex_revalidate.config
     1. Perform any special processing. See [Special Processing](#special-processing).
     1. If a file exists at the path of the file, load it from disk and compare the two.
     1. If there are no changes, don't apply the new file.
     1. If there are changes, backup the existing file in the temp directory, and write the new file.
 1. If configuration was changed which requires an ATS reload to apply, perform a service reload of ATS.
-1. If configuration was changed which requires an ATS restart to apply, and `t3c` is in badass mode, perform a service restart of ATS.
-1. If a sysctl.conf config file was changed, and `t3c` is in badass mode, run `sysctl -p`.
-1. If a ntpd.conf config file was changed, and `t3c` is in badass mode, perform a service restart of ntpd.
+1. If configuration was changed which requires an ATS restart to apply, and `t3c-apply` is in badass mode, perform a service restart of ATS.
+1. If a sysctl.conf config file was changed, and `t3c-apply` is in badass mode, run `sysctl -p`.
+1. If a ntpd.conf config file was changed, and `t3c-apply` is in badass mode, perform a service restart of ntpd.
 1. Update Traffic Ops to unset the Update Pending or Revalidate Pending flag of this Server.
 
 # Special Processing
@@ -90,7 +90,7 @@ Certain config files perform extra processing.
 
 ## Global replacements
 
-All config files have certain text directives replaced. This is done by the atstccfg config generator before the file is returned to `t3c`.
+All config files have certain text directives replaced. This is done by the atstccfg config generator before the file is returned to `t3c-apply`.
 
 * `__SERVER_TCP_PORT__` is replaced with the Server's Port from Traffic Ops; unless the server's port is 80, 0, or null, in which case any occurrences preceded by a colon are removed.
 * `__CACHE_IPV4__` is replaced with the Server's IP address from Traffic Ops.
@@ -100,7 +100,7 @@ All config files have certain text directives replaced. This is done by the atst
 
 ## remap.config
 
-The `t3c` app processes `##OVERRIDE##` directives in the remap.config file.
+The `t3c-apply` app processes `##OVERRIDE##` directives in the remap.config file.
 
 The ##OVERRIDE## template string allows the Delivery Service Raw Remap Text field to override to fully override the Delivery Service’s line in the remap.config ATS configuration file, generated by Traffic Ops. The end result is the original, generated line commented out, prepended with ##OVERRIDDEN## and the ##OVERRIDE## rule is activated in its place. This behavior is used to incrementally deploy plugins used in this configuration file. Normally, this entails cloning the Delivery Service that will have the plugin, ensuring it is assigned to a subset of the cache servers that serve the Delivery Service content, then using this ##OVERRIDE## rule to create a remap.config rule that will use the plugin, overriding the normal rule. Simply grow the subset over time at the desired rate to slowly deploy the plugin. When it encompasses all cache servers that serve the original Delivery Service’s content, the “override Delivery Service” can be deleted and the original can use a non-##OVERRIDE## Raw Remap Text to add the plugin.
 
@@ -108,9 +108,9 @@ The ##OVERRIDE## template string allows the Delivery Service Raw Remap Text fiel
 
 This is presumed to be a udev file for devices which are block devices to be used as disk storage by ATS.
 
-The `t3c` app verifies all devices in the file are owned by the owner listed in the file, and logs errors otherwise.
+The `t3c-apply` app verifies all devices in the file are owned by the owner listed in the file, and logs errors otherwise.
 
-The `t3c` app verifies all devices in the file do not have filesystems. If any device has a filesystem, `t3c` assumes it was a mistake to assign as an ATS storage device, and logs a fatal error.
+The `t3c-apply` app verifies all devices in the file do not have filesystems. If any device has a filesystem, `t3c-apply` assumes it was a mistake to assign as an ATS storage device, and logs a fatal error.
 
 # Trivia
